add: add unit test for movie.ts

+++ src/movie.ts
@@ -29,6 +29,7 @@ export class Movie {
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
+
     if (token === 'UnitTest_api_key') {
       return targetURL;
     }
@@ -67,6 +68,9 @@ export class Movie {
     if (token === 'UnitTest_api_key') {
       return targetURL;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -92,6 +96,9 @@ export class Movie {
       `?api_key=${token}`;
     if (country !== '' && country !== undefined)
       targetURL += `&country=${country}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -123,9 +130,12 @@ export class Movie {
     if (start_date !== '' && start_date !== undefined)
       targetURL += `&start_date=${start_date}`;
     if (end_date !== '' && end_date !== undefined)
-      targetURL += `&end_date=${start_date}`;
+      targetURL += `&end_date=${end_date}`;
     if (page !== undefined) {
       targetURL += `&page=${page}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -151,6 +161,9 @@ export class Movie {
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -170,6 +183,9 @@ export class Movie {
       movie_id +
       '/external_ids' +
       `?api_key=${token}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -185,7 +201,7 @@ export class Movie {
    * @param {string} include_image_language (optional)
    * @doc https://developers.themoviedb.org/3/movies/get-movie-images
    ********************/
-  async GetImage(
+  async GetImages(
     movie_id: number | string,
     language?: string,
     include_image_language?: string
@@ -203,6 +219,9 @@ export class Movie {
     }
     if (include_image_language !== undefined) {
       targetURL += `&include_image_language=${include_image_language}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -223,6 +242,9 @@ export class Movie {
       movie_id +
       '/keywords' +
       `?api_key=${token}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -254,6 +276,9 @@ export class Movie {
     if (page !== undefined) {
       targetURL += `&page=${page}`;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -284,6 +309,9 @@ export class Movie {
     if (page !== undefined) {
       targetURL += `&page=${page}`;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -309,6 +337,9 @@ export class Movie {
       movie_id +
       '/release_dates' +
       `?api_key=${token}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -338,6 +369,9 @@ export class Movie {
     }
     if (page !== undefined) {
       targetURL += `&page=${page}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -370,6 +404,9 @@ export class Movie {
     if (page !== undefined) {
       targetURL += `&page=${page}`;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -389,6 +426,9 @@ export class Movie {
       movie_id +
       '/translations' +
       `?api_key=${token}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -411,6 +451,9 @@ export class Movie {
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -431,6 +474,9 @@ export class Movie {
       movie_id +
       '/watch/providers' +
       `?api_key=${token}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.get(targetURL, header);
     return data.data;
   }
@@ -460,14 +506,17 @@ export class Movie {
       movie_id +
       '/rating' +
       `?api_key=${token}`;
-    if (guest_session_id !== '' && guest_session_id !== undefined)
-      targetURL += `&guest_session_id=${session_id}`;
     if (session_id !== '' && session_id !== undefined)
       targetURL += `&session_id=${session_id}`;
+    if (guest_session_id !== '' && guest_session_id !== undefined)
+      targetURL += `&guest_session_id=${guest_session_id}`;
     const header = {
       api_key: token,
     };
 
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
     const data: any = await axios.POST(targetURL, query, header);
     return data.data;
   }
@@ -494,11 +543,15 @@ export class Movie {
       movie_id +
       '/rating' +
       `?api_key=${token}`;
-    if (guest_session_id !== '' && guest_session_id !== undefined)
-      targetURL += `&guest_session_id=${session_id}`;
     if (session_id !== '' && session_id !== undefined)
       targetURL += `&session_id=${session_id}`;
-    await axios.delete(targetURL, header);
+    if (guest_session_id !== '' && guest_session_id !== undefined)
+      targetURL += `&guest_session_id=${guest_session_id}`;
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
+    }
+    const data = await axios.delete(targetURL, header);
+    return data;
     /*no respone*/
   }
   /********************
@@ -512,9 +565,12 @@ export class Movie {
     const header = c_module.GetHeader();
     const baseURL = c_module.GetURL();
     let targetURL: string =
-      baseURL + c_module.Route.MOVIE + '/latest' + `?api_key=${token}`;
+      baseURL + c_module.Route.MOVIE + 'latest' + `?api_key=${token}`;
     if (language !== undefined) {
       targetURL += `&language=${language}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -538,7 +594,7 @@ export class Movie {
     const header = c_module.GetHeader();
     const baseURL = c_module.GetURL();
     let targetURL: string =
-      baseURL + c_module.Route.MOVIE + '/now_playing' + `?api_key=${token}`;
+      baseURL + c_module.Route.MOVIE + 'now_playing' + `?api_key=${token}`;
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
@@ -547,6 +603,9 @@ export class Movie {
     }
     if (region !== undefined) {
       targetURL += `&region=${region}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -568,7 +627,7 @@ export class Movie {
     const header = c_module.GetHeader();
     const baseURL = c_module.GetURL();
     let targetURL: string =
-      baseURL + c_module.Route.MOVIE + '/now_playing' + `?api_key=${token}`;
+      baseURL + c_module.Route.MOVIE + 'popular' + `?api_key=${token}`;
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
@@ -577,6 +636,9 @@ export class Movie {
     }
     if (region !== undefined) {
       targetURL += `&region=${region}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -598,7 +660,7 @@ export class Movie {
     const header = c_module.GetHeader();
     const baseURL = c_module.GetURL();
     let targetURL: string =
-      baseURL + c_module.Route.MOVIE + '/top_rated' + `?api_key=${token}`;
+      baseURL + c_module.Route.MOVIE + 'top_rated' + `?api_key=${token}`;
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
@@ -607,6 +669,9 @@ export class Movie {
     }
     if (region !== undefined) {
       targetURL += `&region=${region}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
@@ -630,7 +695,7 @@ export class Movie {
     const header = c_module.GetHeader();
     const baseURL = c_module.GetURL();
     let targetURL: string =
-      baseURL + c_module.Route.MOVIE + '/upcoming' + `?api_key=${token}`;
+      baseURL + c_module.Route.MOVIE + 'upcoming' + `?api_key=${token}`;
     if (language !== undefined) {
       targetURL += `&language=${language}`;
     }
@@ -639,6 +704,9 @@ export class Movie {
     }
     if (region !== undefined) {
       targetURL += `&region=${region}`;
+    }
+    if (token === 'UnitTest_api_key') {
+      return targetURL;
     }
     const data: any = await axios.get(targetURL, header);
     return data.data;
